// src/journey/details/component/journeyDetailsCard/insurance/types/InsuranceTypes.ts
export interface TvlG {
  tvlA?: string[];
}

export interface InsuranceApiData {
  bid?: string;
  idx?: number;
  nm?: string;
  cdnm?: string;
  ttl?: string;
  title?: string;
  typ?: string;
  typD?: string;
  paxD?: string;
  tvlG?: TvlG;
  dploc?: string;
  pkloc?: string;
  dsc?: string;
  ntA?: string[];
  isInc?: boolean; 
}